positionable: further minor refactor
 - move the header under model/, it's not used by the GUI code
 - add initializers and use default constructors for Positionable and ParamNode
